XmlStructWriter.write_xml — Method

write_xml(xml_object, xml_path::AbstractString)::Nothing

Write the given xml_object to the path specified by xml_path.

The xml_object must be of a type generated by the XsdToStruct package. It is also necessary that the dict in the __xml_attributes field of this object has the following two keys:

A key __root_name that should contain the name of the root element of the xml file.

An entry with the correct xmlns attribute that gives name of the namespace and the corresponding URI of the root element, see the xml specification for details.

Please consult the given examples if this is unclear; or inspect an XML loaded with XmlStructLoader, it will contain all the needed entries in the __xml_attributes field.

With this input the written XML is valid according to the schema that was used to generate the struct definitions with XsdToStruct.

Examples

Creating an XML from scratch:

julia> include(joinpath("Example", "Example.jl"))
julia> using .ExampleSchema

julia> using XmlStructWriter

julia> Example1 = ExampleComplexType(
	Element_double=100.22,
	Element_boolean=true,
	Element_integer=-778
)

julia> xml_attributes = Dict(
	"xmlns:xsi"=>"http://www.w3.org/2001/XMLSchema-instance",
	"xsi:schemaLocation"=>"ExampleSchema Example.xsd",
	"xmlns:ExampleSchema"=>"ExampleSchema",
	"__root_name"=>"document"
)

julia> document = documentType(
	Example1=Example1,
	Example2="AA99",
	__xml_attributes=xml_attributes
)

julia> write_xml(document, "example.xml")

Loading an XML, editing a field, and then writing to a new file, we are using Accessors.jl here which is recommended:

julia> include(joinpath("Example", "Example.jl"))
julia> using .ExampleSchema

julia> using XmlStructLoader
julia> using XmlStructWriter
julia> using Accessors

julia> xml_original = load("example.xml", ExampleSchema)

julia> xml_new = @set xml_original.Example1.Element_double = -789.123
julia> write_xml(xml_new, "example_new.xml")