AbstractXsdTypes.AbstractXSDComplex — Type

AbstractXSDComplex

Abstract type used for complex XSD structs.

The contract for concrete subtypes is:

each fieldname is an XML element

except __xml_attributes, which contains the attributes

except __validated, which indicates if the object was validated at construction with respect to the XSD schema

This type has a special Base.show which calls AbstractXsdTypes.print_tree

import AbstractXsdTypes: AbstractXSDComplex

struct SimpleType <: AbstractXSDComplex
    Element1::String
    Element2::Float64
    __xml_attributes::Union{Nothing,Dict{String,String}}
    __validated::Bool
end

simple = SimpleType("value", 1e-3, Dict("x" => "y"), true)
show(simple)

# output

|-- Element1: value
|-- Element2: 0.001
|-- __xml_attributes: Dict("x" => "y")
|-- __validated: true

AbstractXsdTypes.AbstractXSDFloat — Type

AbstractXSDFloat <: AbstractFloat

Must have a value field, an __xml_attributes field, and a __validated field.

import AbstractXsdTypes: AbstractXSDFloat

struct FloatType <: AbstractXSDFloat
    value::Float64
    __xml_attributes::Union{Nothing,Dict{String,String}}
    __validated::Bool
end

xsd_float = FloatType(1.0, Dict("x" => "y"), true)
show(xsd_float)

# output

1.0

AbstractXsdTypes.AbstractXSDSigned — Type

AbstractXSDSigned <: Signed

Must have a value field, an __xml_attributes field, and a __validated field.

import AbstractXsdTypes: AbstractXSDSigned

struct SignedType <: AbstractXSDSigned
    value::Int64
    __xml_attributes::Union{Nothing,Dict{String,String}}
    __validated::Bool
end

xsd_signed = SignedType(1, Dict("x" => "y"), true)
show(xsd_signed)

# output

1

AbstractXsdTypes.AbstractXSDString — Type

AbstractXSDString <: AbstractString

Must have a value field, an __xml_attributes field, and a __validated field.

import AbstractXsdTypes: AbstractXSDString

struct StringType <: AbstractXSDString
    value::String
    __xml_attributes::Union{Nothing,Dict{String,String}}
    __validated::Bool
end

xsd_string = StringType("value", Dict("x" => "y"), true)
show(xsd_string)

# output

"value"

AbstractXsdTypes.AbstractXSDUnion — Type

AbstractXsdTypes.AbstractXSDUnion

Abstract type used for XSD Union types.

First field must be the value, which will be another XSD type.

You must use the AbstractXsdTypes.union_value function in the constructor as shown below. This is to make sure the values get converted to a union value.

You must implement the AbstractXsdTypes.union_types function.

xml attributes will be passed along to the value type.

Base.@kwdef @concrete struct UnionType <: AbstractXsdTypes.AbstractXSDUnion
    value <: Union{FloatType,AnotherFloatType,StringType}
    __validated::Bool = true
end

AbstractXsdTypes.union_types(::Type{<:UnionType}) = (FloatType, AnotherFloatType, StringType)

AbstractXsdTypes.AbstractXSDUnsigned — Type

AbstractXSDUnsigned <: Unsigned

Must have a value field, an __xml_attributes field, and a __validated field.

import AbstractXsdTypes: AbstractXSDUnsigned

struct UnsignedType <: AbstractXSDUnsigned
    value::UInt64
    __xml_attributes::Union{Nothing,Dict{String,String}}
    __validated::Bool
end

xsd_unsigned = UnsignedType(0x0000000000000001, Dict("x" => "y"), true)
show(xsd_unsigned)

# output

0x0000000000000001

AbstractXsdTypes.XSDRestrictionViolationError — Type

XSDRestrictionViolationError <: Exception

Custom exception type to indicate that something was passed to an XSD type constructor which violates a restriction for that type.

AbstractXsdTypes.XSDStringRestrictionViolationError — Type

XSDStringRestrictionViolationError <: XSDRestrictionViolationError

Custom exception type to indicate that a string was passed to an XSD type constructor which violates a string based restriction for that type.

AbstractXsdTypes.XSDValueRestrictionViolationError — Type

XSDValueRestrictionViolationError <: XSDRestrictionViolationError

Custom exception type to indicate that a value was passed to an XSD type constructor which violates a value based restriction for that type.

AbstractXsdTypes.can_be_converted — Method

can_be_converted(::Type{T}, ::Type{S})::Bool where {T<:AbstractXSDComplex, S<:AbstractXSDComplex}

Naive check to see if type S can be converted to type T. This function will only look for compatible field names, compatibility of field types is not checked.

AbstractXsdTypes.defaults — Function

AbstractXsdTypes.defaults(::Type{MyType}) = (:field_a=default_value, :field_b=>default_value)

Define default arguments for various fields of MyType, which will be generated by XsdToStruct and can be used by your custom XML parser.

AbstractXsdTypes.print_tree — Method

print_treeprint_tree(
    io::IO,
    x::AbstractXSDComplex;
    print_all::Bool=false,
    indent_string::AbstractString=""
)::Nothing

Prints the AbstractXSDComplex as a tree.