XsdToStruct.generate_modules — Method

generate_modules(xsd_locations::Dict{AbstractString, AbstractString}, xsd_modules_path::AbstractString)::Nothing

Generate modules from the given locations. The xsd_locations should be a dict which maps xsd names to their locations. The location can either be a path to a file, or a directory where the xsd file is located, or a URL where the xsd file can downloaded from. The generated modules are placed into the given xsd_modules_path. It is recomended for any packages that use generated modules to use this function in a build script to generate up to date modules on the fly.

Examples

Let's assume we have a folder "xsdfiles" which contains the xsd files: "file1.xsd", "file2.xsd", and "file3.xsd", and we also have a file located at a URL. We can then generate modules for them and place these modules into the folder "xsd_modules" as follows:

julia> using XsdToStruct
julia> xsd_locations = Dict(
    "local_file_1" => joinpath("path", "to", "xsd_files", "file_1.xsd"),
    "local_file_2" => joinpath("path", "to", "xsd_files"),
    "local_file_3" => joinpath("path", "to", "xsd_files"),
    "remote_file" => "https://url.to.last/xsd/file"
)
julia> xsd_modules_path = joinpath("output", "path", "for", "xsd", "modules")
julia> generate_modules(xsd_locations, xsd_modules_path)

XsdToStruct.get_field_indices_of_type — Method

get_field_indices_of_type(
    node::Union{ComplexTreeNode, SchemaTreeNode},
    ::Type{T}
) where T <: AbstractFieldData

Find indices for all elements in the fields attribute of the given node which are of the given FieldData type. Note that this does not include the child fields in the case of a ComplexTreeNode.

XsdToStruct.parse_xsd_element_child — Method

parse_xsd_element_child(
    xsd_element::XMLElement,
    parent_name::AbstractString
)::Tuple{FieldData, AbstractTreeNode}

Extract child node from xsd element with a type defined inside the element

XsdToStruct.qualified_type — Method

qualified_type(
    field_data::Union{FieldData, ChoiceFieldData},
    xsd_namespace::AbstractString
)::String

Return the qualified type name for the given field_data relative to the namespace defined by the xsd. Hence if any leading xsd namespace part is present in the qualified type name it will be removed.

XsdToStruct.qualified_type — Method

qualified_type(
    field_data::Union{FieldData, ChoiceFieldData}
)::String

Return the qualified type name for the given field_data as read from the xsd. This will include the xsd namespace explicitly if it was given in the xsd. If the qualified name relative to the xsd namespace is needed add the xsd namespace as a string as the second argument.

XsdToStruct.qualified_type — Method

qualified_type(field_data::GroupFieldData)::String

An Xsd group does not really have a qualified name, hence just the name is returned

XsdToStruct.required_julia_types — Method

required_julia_types(complex_node::ComplexTreeNode)::Vector{String}

Returns types which need to be defined in order for the struct definition of 'complex_node' to be valid.

XsdToStruct.xsd_to_struct_module — Method

xsd_to_struct_module(xsd_path::AbstractString, output_dir::AbstractString)::String

Generate Julia module corresponding to the given xsd file. The resulting Julia module will be placed in the given output_dir. The module generated will have the same name as the xsd namespace defined in the given xsd file. The path to the file with the generated code is given as a return value. The file generated will have the approximately the same name as the given xsd file but with a ".jl" extension instead of ".xsd". To use the generated module include the generated file that is returned by this function, and then use or import the module.

Examples

julia> using XsdToStruct
julia> xsd_to_struct_module(joinpath("path", "to", "file.xsd"), joinpath("output", "dir"))
julia> include(joinpath("output", "dir", "file", "file.jl"))
julia> using file_xsd_namespace

or

julia> using XsdToStruct
julia> xsd_to_struct_module(joinpath("path", "to", "file.xsd"), joinpath("output", "dir"))
julia> include(joinpath("output", "dir", "file", "file.jl"))
julia> import file_xsd_namespace

XsdToStruct.xsd_to_struct_module — Method

xsd_to_struct_module(xsd_path::AbstractString)::String

Generate Julia module corresponding to the given xsd file. The resulting Julia module will be placed in the same directory as the given xsd file. The path to the file with the generated code is given as a return value.

Examples

julia> using XsdToStruct
julia> xsd_to_struct_module(joinpath("path", "to", "file.xsd"))
julia> include(joinpath("path", "to", "file", "file.jl"))
julia> using file_xsd_namespace

or

julia> using XsdToStruct
julia> xsd_to_struct_module(joinpath("path", "to", "file.xsd"))
julia> include(joinpath("path", "to", "file", "file.jl"))
julia> import file_xsd_namespace